Sharded 3pc implementation finished. Starting final testing in docker environemnt.

+++ platform/consensus/ordering/sharded_3pc/shard_communicator.h
@@ -11,7 +11,7 @@
 namespace resdb {
 
 /**************************
- * ShardCommunicator provides an abstraction layer for sending messages to specific nodes, shards, or shard leaders based on the shard metadata. 
+ * ShardCommunicator provides an abstraction layer for sending messages to specific nodes, shards, or shard leaders based on the shard metadata.
  * It uses the underlying ReplicaCommunicator to send messages to the appropriate recipients as determined by the ShardMetadata.
  * - replica_communicator: instance of ReplicaCommunicator.
  * - shard_metadata: metadata object containing shard configuration and node-shard mappings.
@@ -24,7 +24,7 @@ class ShardCommunicator {
                     const std::vector<ReplicaInfo>& replicas);
 
   // Methods to send messages to specific nodes, shards, or shard leaders.
-  int SendToNode(const google::protobuf::Message& message, uint32_t node_id);
+  int SendToNode(const google::protobuf::Message& message, int64_t node_id);
   int BroadcastToShard(const google::protobuf::Message& message,
                        uint32_t shard_id, bool include_self = true);
   int BroadcastToLocalShard(const google::protobuf::Message& message,
@@ -36,13 +36,13 @@ class ShardCommunicator {
 
  private:
   // Helper methods to get ReplicaInfo for a given node ID and to send messages to a list of node IDs.
-  const ReplicaInfo& ReplicaForNode(uint32_t node_id) const;
+  const ReplicaInfo& ReplicaForNode(int64_t node_id) const;
   int SendToNodes(const google::protobuf::Message& message,
-                  const std::vector<uint32_t>& node_ids, bool include_self);
+                  const std::vector<int64_t>& node_ids, bool include_self);
 
-  ReplicaCommunicator* replica_communicator_;
-  const ShardMetadata* shard_metadata_;
-  std::map<uint32_t, ReplicaInfo> replicas_by_node_id_;
+  ReplicaCommunicator* replica_communicator_;             // Underlying communicator for sending messages.
+  const ShardMetadata* shard_metadata_;                   // Metadata containing shard configuration and node-shard mappings.
+  std::map<int64_t, ReplicaInfo> replicas_by_node_id_;    // Mapping of node IDs to their corresponding ReplicaInfo for quick lookup when sending messages.
 };
 
 }  // namespace resdb
